moved VoxelMap to it's own module and fixed issue #5
